refactor: move columnsToShow to logic to custom hook
