+++ src/presentation/react/core/components/metadata-table/MetadataTable.tsx
@@ -149,9 +149,7 @@ const MetadataTable: React.FC<MetadataTableProps> = ({
 }) => {
     const { compositionRoot, api: defaultApi } = useAppContext();
     const classes = useStyles();
-
     const snackbar = useSnackbar();
-    const { visibleColumns, saveReorderedColumns } = useTableColumns(Namespace.METADATA_USER_COLUMNS);
 
     const [model, updateModel] = useState<typeof D2Model>(() => models[0] ?? DataElementModel);
     const [ids, updateIds] = useState<string[]>([]);
@@ -701,6 +699,8 @@ const MetadataTable: React.FC<MetadataTableProps> = ({
         ])
     );
 
+    const { columnsToShow, saveReorderedColumns } = useTableColumns(Namespace.METADATA_USER_COLUMNS, columns);
+
     const details: ObjectsTableDetailField<MetadataType>[] = uniqCombine([...model.getDetails(), responsibleField]);
 
     const actions: TableAction<MetadataType>[] = uniqCombine([...tableActions, ...additionalActions]);
@@ -710,20 +710,6 @@ const MetadataTable: React.FC<MetadataTableProps> = ({
     }, [childrenSelection, selection]);
 
     const shownRows = useMemo(() => (modelIsSyncAll ? [] : transformRows(rows)), [modelIsSyncAll, rows, transformRows]);
-
-    const columnsToShow = useMemo(() => {
-        if (!visibleColumns || _.isEmpty(visibleColumns)) return columns;
-
-        const indexes = _(visibleColumns)
-            .map((columnName, idx) => [columnName, idx] as [string, number])
-            .fromPairs()
-            .value();
-
-        return _(columns)
-            .map(column => ({ ...column, hidden: !visibleColumns.includes(column.name as string) }))
-            .sortBy(column => indexes[column.name] || 0)
-            .value();
-    }, [visibleColumns, columns]);
 
     return (
         <React.Fragment>
